+++ pineappl_cli/src/convolve.rs
@@ -58,7 +58,7 @@ pub struct Opts {
 impl Subcommand for Opts {
     fn run(&self, cfg: &GlobalConfiguration) -> Result<ExitCode> {
         let grid = helpers::read_grid(&self.input)?;
-        let mut conv_funs_0 = helpers::create_conv_funs(&self.conv_funs[0])?;
+        let mut conv_funs_0 = helpers::create_conv_funs(&self.conv_funs[0], cfg.pdf_backend)?;
         let bins: Vec<_> = self.bins.iter().cloned().flatten().collect();
 
         let results = helpers::convolve_scales(
@@ -91,7 +91,7 @@ impl Subcommand for Opts {
             .iter()
             .flat_map(|conv_funs| {
                 let conv_types = &conv_funs.conv_types;
-                let mut conv_funs = helpers::create_conv_funs(conv_funs).unwrap();
+                let mut conv_funs = helpers::create_conv_funs(conv_funs, cfg.pdf_backend).unwrap();
                 helpers::convolve(
                     &grid,
                     &mut conv_funs,