+++ pineappl_cli/src/plot.rs
@@ -63,7 +63,8 @@ impl Subcommand for Opts {
         };
 
         let grid = helpers::read_grid(&self.input)?;
-        let mut conv_funs = helpers::create_conv_funs(&self.conv_funs[0])?;
+        let mut conv_funs =
+            helpers::create_conv_funs(&self.conv_funs[0], cfg.pdf_backend)?;
         let slices = grid.bwfl().slices();
         let mut data_string = String::new();
 
@@ -159,7 +160,8 @@ impl Subcommand for Opts {
                 .map(|conv_funs| {
                     if self.no_conv_fun_unc {
                         let conv_types = &conv_funs.conv_types;
-                        let mut conv_funs = helpers::create_conv_funs(conv_funs)?;
+                        let mut conv_funs =
+                            helpers::create_conv_funs(conv_funs, cfg.pdf_backend)?;
 
                         let results = helpers::convolve(
                             &grid,
@@ -178,6 +180,7 @@ impl Subcommand for Opts {
                         let (set, funs) = helpers::create_conv_funs_for_set(
                             conv_funs,
                             self.conv_fun_uncert_from,
+                            cfg.pdf_backend,
                         )?;
 
                         let pdf_results: Vec<_> = funs